moved menu and enum to Utility folder, made for anything that can be used as generic help methods etc
