+++ GameService.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace LordsOfArda
+{
+    public class GameService
+    {
+        public void StartGame(Character character)
+        {
+            bool isContinue = true;
+            while (isContinue)
+            {
+                
+            }
+        }
+    }
+}
