+++ Utility/Menu.cs
@@ -6,7 +6,7 @@ using System.Security.Cryptography.X509Certificates;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace LordsOfArda
+namespace LordsOfArda.Utility
 {
     internal static class Menu
     {
